+++ openpaw/runtime/scheduling/__init__.py
@@ -4,7 +4,21 @@ Provides cron and heartbeat scheduling capabilities.
 """
 
 from openpaw.runtime.scheduling.cron import CronScheduler
+from openpaw.runtime.scheduling.cron_executor import CronExecutor
+from openpaw.runtime.scheduling.cron_job_manager import CronJobManager
 from openpaw.runtime.scheduling.heartbeat import HeartbeatScheduler
+from openpaw.runtime.scheduling.heartbeat_executor import HeartbeatExecutor
+from openpaw.runtime.scheduling.heartbeat_preflight import HeartbeatPreflight
+from openpaw.runtime.scheduling.heartbeat_prompt import HeartbeatPromptBuilder
 from openpaw.runtime.scheduling.loader import CronLoader
 
-__all__ = ["CronScheduler", "HeartbeatScheduler", "CronLoader"]
+__all__ = [
+    "CronExecutor",
+    "CronJobManager",
+    "CronLoader",
+    "CronScheduler",
+    "HeartbeatExecutor",
+    "HeartbeatPreflight",
+    "HeartbeatPromptBuilder",
+    "HeartbeatScheduler",
+]
