+++ openpaw/runtime/scheduling/heartbeat.py
@@ -1,39 +1,20 @@
 """HeartbeatScheduler for periodic agent task evaluation."""
 
-import json
 import logging
 from collections.abc import Awaitable, Callable, Mapping
-from datetime import UTC, datetime, time
 from pathlib import Path
-from typing import TYPE_CHECKING, Any
+from typing import Any
 
-import yaml
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from apscheduler.triggers.interval import IntervalTrigger
 
 from openpaw.agent.session_logger import SessionLogger
-from openpaw.builtins.tools._channel_context import (
-    clear_channel_context,
-    set_channel_context,
-    set_invocation_origin,
-)
-from openpaw.channels.base import ChannelAdapter
 from openpaw.core.config import HeartbeatConfig
-from openpaw.core.paths import HEARTBEAT_LOG_JSONL, HEARTBEAT_MD, TASKS_YAML
-from openpaw.core.prompts.heartbeat import (
-    ACTIVE_TASKS_TEMPLATE,
-    HEARTBEAT_PROMPT,
-    build_task_summary,
-)
-from openpaw.core.prompts.system_events import (
-    HEARTBEAT_RESULT_TEMPLATE,
-    HEARTBEAT_RESULT_TRUNCATED_TEMPLATE,
-    INJECTION_TRUNCATION_LIMIT,
-)
+from openpaw.core.prompts.heartbeat import HEARTBEAT_PROMPT  # noqa: F401
 from openpaw.core.timezone import workspace_now
-
-if TYPE_CHECKING:
-    pass
+from openpaw.runtime.scheduling.heartbeat_executor import HeartbeatExecutor
+from openpaw.runtime.scheduling.heartbeat_preflight import HeartbeatPreflight
+from openpaw.runtime.scheduling.heartbeat_prompt import HeartbeatPromptBuilder
 
 logger = logging.getLogger(__name__)
 
@@ -51,7 +32,7 @@ class HeartbeatScheduler:
         workspace_name: str,
         workspace_path: Path,
         agent_factory: Callable[..., Any],
-        channels: Mapping[str, ChannelAdapter],
+        channels: Mapping[str, Any],
         config: HeartbeatConfig,
         timezone: str = "UTC",
         token_logger: Any | None = None,
@@ -59,22 +40,7 @@ class HeartbeatScheduler:
         session_logger: SessionLogger | None = None,
         ack_tool: Any | None = None,
     ):
-        """Initialize the heartbeat scheduler.
-
-        Args:
-            workspace_name: Name of the agent workspace.
-            workspace_path: Path to the workspace directory.
-            agent_factory: Factory function to create fresh agent instances (no checkpointer).
-            channels: Dictionary mapping channel types to channel instances for routing.
-            config: Heartbeat configuration settings.
-            timezone: IANA timezone identifier for workspace timezone (default: "UTC").
-            token_logger: Optional TokenUsageLogger for logging token metrics.
-            result_callback: Optional callback for queue injection of results.
-            session_logger: Optional SessionLogger for writing session logs.
-            ack_tool: Optional AcknowledgeTool instance for unified silence. When the
-                heartbeat agent calls acknowledge_event(), this takes precedence over
-                the HEARTBEAT_OK magic string.
-        """
+        """Initialize the heartbeat scheduler."""
         self.workspace_name = workspace_name
         self.workspace_path = workspace_path
         self.agent_factory = agent_factory
@@ -88,158 +54,49 @@ class HeartbeatScheduler:
         self._scheduler: AsyncIOScheduler | None = None
         self._job: Any = None
 
-        # Parse active hours at initialization
-        self._active_hours = self._parse_active_hours(config.active_hours)
+        self._preflight = HeartbeatPreflight(workspace_path, timezone)
+        self._prompt_builder = HeartbeatPromptBuilder()
+        self._active_hours = self._preflight.parse_active_hours(config.active_hours)
+        self._executor = HeartbeatExecutor(
+            workspace_path=workspace_path,
+            workspace_name=workspace_name,
+            config=config,
+            token_logger=token_logger,
+            result_callback=result_callback,
+            session_logger=session_logger,
+            ack_tool=ack_tool,
+        )
 
     @staticmethod
-    def _resolve_heartbeat_session_key(channel: ChannelAdapter, config: HeartbeatConfig) -> str | None:
-        """Resolve session key from heartbeat config.
+    def _resolve_heartbeat_session_key(channel: Any, config: HeartbeatConfig) -> str | None:
+        """Resolve session key from heartbeat config."""
+        return HeartbeatExecutor._resolve_heartbeat_session_key(channel, config)
 
-        Uses ``target_id`` (preferred) with fallback to legacy ``target_chat_id``/``target_channel_id``.
-
-        Args:
-            channel: The channel adapter instance to build the session key against.
-            config: The heartbeat configuration specifying channel name and target ID.
-
-        Returns:
-            A session key string, or None if no supported routing config is found.
-        """
-        target = next(
-            (v for v in (config.target_id, config.target_chat_id, config.target_channel_id) if v is not None),
-            None,
-        )
-        if target:
-            return channel.build_session_key(target)
-        return None
-
-    def _parse_active_hours(self, active_hours: str | None) -> tuple[time, time] | None:
-        """Parse active hours string like '08:00-22:00' into start/end times.
-
-        Args:
-            active_hours: String in format "HH:MM-HH:MM" or None.
-
-        Returns:
-            Tuple of (start_time, end_time) or None if always active.
-
-        Raises:
-            ValueError: If the format is invalid.
-        """
-        if not active_hours:
-            return None
-
-        try:
-            start_str, end_str = active_hours.split("-")
-            start_hour, start_min = map(int, start_str.strip().split(":"))
-            end_hour, end_min = map(int, end_str.strip().split(":"))
-
-            start_time = time(start_hour, start_min)
-            end_time = time(end_hour, end_min)
-
-            return (start_time, end_time)
-        except (ValueError, AttributeError) as e:
-            raise ValueError(f"Invalid active_hours format: {active_hours}. Expected 'HH:MM-HH:MM'") from e
+    def _parse_active_hours(self, active_hours: str | None) -> Any:
+        """Parse active hours string like '08:00-22:00' into start/end times."""
+        return self._preflight.parse_active_hours(active_hours)
 
     def _is_within_active_hours(self) -> bool:
-        """Check if current time is within active hours window.
-
-        Returns:
-            True if within active hours or if no active hours are set (always active).
-        """
-        if self._active_hours is None:
-            return True  # Always active if no hours specified
-
+        """Check if current time is within active hours window."""
         current_time = workspace_now(self._timezone).time()
-        start_time, end_time = self._active_hours
-
-        # Handle case where active hours span midnight
-        if start_time <= end_time:
-            # Normal case: 08:00-22:00
-            return start_time <= current_time <= end_time
-        else:
-            # Midnight span: 22:00-08:00
-            return current_time >= start_time or current_time <= end_time
+        return self._preflight.is_within_active_hours(self._active_hours, current_time)
 
     def _is_heartbeat_ok(self, response: str) -> bool:
-        """Check if response indicates no action needed.
-
-        Args:
-            response: Agent response text.
-
-        Returns:
-            True if response contains HEARTBEAT_OK (case-insensitive).
-        """
-        return "HEARTBEAT_OK" in response.upper()
+        """Check if response indicates no action needed."""
+        return self._preflight.is_heartbeat_ok(response)
 
     def _build_task_summary(self, tasks: list[dict[str, Any]]) -> str | None:
-        """Build a compact task summary from TASKS.yaml data.
-
-        Thin wrapper around build_task_summary from prompts.heartbeat.
-
-        Args:
-            tasks: List of task dictionaries (already filtered to active tasks).
-
-        Returns:
-            Formatted task summary string, or None if no tasks.
-        """
-        return build_task_summary(tasks)
+        """Build a compact task summary from TASKS.yaml data."""
+        return self._prompt_builder.build_task_summary(tasks)
 
     def _build_heartbeat_prompt(self, task_summary: str | None = None) -> str:
-        """Build the heartbeat prompt with current timestamp and optional task summary.
-
-        Args:
-            task_summary: Optional compact task summary to inject into the prompt.
-
-        Returns:
-            Formatted heartbeat prompt string.
-        """
+        """Build the heartbeat prompt with current timestamp and optional task summary."""
         timestamp = workspace_now(self._timezone).isoformat()
-        prompt = HEARTBEAT_PROMPT.format(timestamp=timestamp)
-
-        if task_summary:
-            prompt += "\n" + ACTIVE_TASKS_TEMPLATE.format(task_summary=task_summary)
-
-        return prompt
+        return self._prompt_builder.build_heartbeat_prompt(timestamp, task_summary)
 
     def _should_skip_heartbeat(self) -> tuple[bool, str, str | None, int]:
-        """Pre-flight check: skip heartbeat if nothing needs attention.
-
-        Checks HEARTBEAT.md and TASKS.yaml to determine if LLM invocation
-        can be skipped, saving API costs for idle workspaces.
-
-        Returns:
-            Tuple of (should_skip, reason, task_summary, task_count).
-            task_summary is None if skipping or no active tasks, otherwise a formatted string.
-            task_count is the number of active tasks found (0 if none or on error).
-        """
-        heartbeat_md = self.workspace_path / str(HEARTBEAT_MD)
-        heartbeat_empty = True
-        if heartbeat_md.exists():
-            try:
-                content = heartbeat_md.read_text().strip()
-                heartbeat_empty = not content or content == "# Heartbeat" or len(content) < 20
-            except OSError:
-                heartbeat_empty = False  # Can't read = don't skip
-
-        tasks_file = self.workspace_path / str(TASKS_YAML)
-        active_tasks = []
-        if tasks_file.exists():
-            try:
-                with tasks_file.open() as f:
-                    data = yaml.safe_load(f)
-                tasks = data.get("tasks", []) if data else []
-                active_statuses = {"pending", "in_progress", "awaiting_check"}
-                active_tasks = [t for t in tasks if t.get("status") in active_statuses]
-            except (yaml.YAMLError, OSError) as e:
-                logger.warning(f"Failed to read TASKS.yaml during pre-flight: {e}")
-                # Can't read = don't skip, but no task summary
-                return False, "pre-flight checks passed (TASKS.yaml read error)", None, 0
-
-        if heartbeat_empty and not active_tasks:
-            return True, "no active tasks and HEARTBEAT.md is empty", None, 0
-
-        # Build task summary if we're not skipping and have active tasks
-        task_summary = self._build_task_summary(active_tasks) if active_tasks else None
-        return False, "pre-flight checks passed", task_summary, len(active_tasks)
+        """Pre-flight check: skip heartbeat if nothing needs attention."""
+        return self._preflight.should_skip_heartbeat()
 
     def _record_heartbeat_event(
         self,
@@ -255,53 +112,20 @@ class HeartbeatScheduler:
         response: str | None = None,
         tools_used: list[str] | None = None,
     ) -> None:
-        """Append heartbeat event to workspace JSONL log.
-
-        Args:
-            outcome: Event outcome (ran, skipped, heartbeat_ok, error).
-            reason: Reason for skip or additional context.
-            duration_ms: Execution duration in milliseconds.
-            error: Error message if applicable.
-            input_tokens: Input token count from invocation.
-            output_tokens: Output token count from invocation.
-            total_tokens: Total token count from invocation.
-            llm_calls: Number of LLM calls made.
-            task_count: Number of active tasks when heartbeat ran.
-            response: Full agent response text.
-            tools_used: List of tool names invoked during the heartbeat.
-        """
-        log_path = self.workspace_path / str(HEARTBEAT_LOG_JSONL)
-        event: dict[str, Any] = {
-            "timestamp": datetime.now(UTC).isoformat(),
-            "workspace": self.workspace_name,
-            "outcome": outcome,
-        }
-        if reason:
-            event["reason"] = reason
-        if duration_ms is not None:
-            event["duration_ms"] = round(duration_ms, 1)
-        if error:
-            event["error"] = error
-        if input_tokens is not None:
-            event["input_tokens"] = input_tokens
-        if output_tokens is not None:
-            event["output_tokens"] = output_tokens
-        if total_tokens is not None:
-            event["total_tokens"] = total_tokens
-        if llm_calls is not None:
-            event["llm_calls"] = llm_calls
-        if task_count is not None:
-            event["task_count"] = task_count
-        if tools_used:
-            event["tools_used"] = tools_used
-        if response:
-            event["response"] = response
-
-        try:
-            with log_path.open("a", encoding="utf-8") as f:
-                f.write(json.dumps(event) + "\n")
-        except OSError as e:
-            logger.warning(f"Failed to write heartbeat log: {e}")
+        """Append heartbeat event to workspace JSONL log."""
+        self._executor.record_heartbeat_event(
+            outcome=outcome,
+            reason=reason,
+            duration_ms=duration_ms,
+            error=error,
+            input_tokens=input_tokens,
+            output_tokens=output_tokens,
+            total_tokens=total_tokens,
+            llm_calls=llm_calls,
+            task_count=task_count,
+            response=response,
+            tools_used=tools_used,
+        )
 
     async def start(self) -> None:
         """Start the heartbeat scheduler with interval trigger."""
@@ -338,15 +162,13 @@ class HeartbeatScheduler:
 
     async def _run_heartbeat(self) -> None:
         """Execute a heartbeat check with pre-flight skip and event logging."""
-        import time as time_module
-
         # Check active hours
         if not self._is_within_active_hours():
             logger.debug(
                 f"Heartbeat skipped for '{self.workspace_name}' "
                 f"(outside active hours: {self.config.active_hours})"
             )
-            self._record_heartbeat_event("skipped", reason="outside active hours")
+            self._executor.record_heartbeat_event("skipped", reason="outside active hours")
             return
 
         # Pre-flight check (returns task count alongside summary)
@@ -354,254 +176,15 @@ class HeartbeatScheduler:
 
         if should_skip:
             logger.info(f"Heartbeat skipped for '{self.workspace_name}': {reason}")
-            self._record_heartbeat_event("skipped", reason=reason, task_count=0)
+            self._executor.record_heartbeat_event("skipped", reason=reason, task_count=0)
             return
 
-        logger.info(f"Running heartbeat check for workspace: {self.workspace_name}")
-        set_invocation_origin(f"heartbeat:{self.workspace_name}")
-
-        # Set up channel context so send_message/send_file work during heartbeat execution
-        heartbeat_channel = self.channels.get(self.config.target_channel)
-        heartbeat_session_key = (
-            self._resolve_heartbeat_session_key(heartbeat_channel, self.config)
-            if heartbeat_channel
-            else None
+        await self._executor.run_heartbeat(
+            preflight=self._preflight,
+            prompt_builder=self._prompt_builder,
+            timezone=self._timezone,
+            task_summary=task_summary,
+            task_count=task_count,
+            agent_factory=self.agent_factory,
+            channels=self.channels,
         )
-        if heartbeat_channel and heartbeat_session_key:
-            set_channel_context(heartbeat_channel, heartbeat_session_key)
-
-        start_time = time_module.monotonic()
-
-        try:
-            # Resolve max_output_tokens: heartbeat config takes precedence over
-            # the workspace-level default already baked into the factory's
-            # extra_model_kwargs. Passing it here as extra_overrides lets the
-            # heartbeat-specific cap win.
-            extra_overrides: dict[str, Any] = {}
-            if self.config.max_output_tokens is not None:
-                extra_overrides["max_tokens"] = self.config.max_output_tokens
-
-            agent_runner = (
-                self.agent_factory(extra_overrides=extra_overrides)
-                if extra_overrides
-                else self.agent_factory()
-            )
-            heartbeat_prompt = self._build_heartbeat_prompt(task_summary=task_summary)
-            response = await agent_runner.run(message=heartbeat_prompt)
-            duration_ms = (time_module.monotonic() - start_time) * 1000
-
-            # Extract metrics and activity from agent runner
-            metrics = agent_runner.last_metrics
-
-            # Log a warning when the output cap was hit — response is likely truncated
-            cap = agent_runner.max_output_tokens
-            if isinstance(cap, int) and metrics and isinstance(metrics.output_tokens, int):
-                if metrics.output_tokens >= cap:
-                    logger.warning(
-                        f"Heartbeat output token cap reached: "
-                        f"{metrics.output_tokens}/{cap} tokens "
-                        f"— response likely truncated (workspace: {self.workspace_name})"
-                    )
-            input_tokens = metrics.input_tokens if metrics else None
-            output_tokens = metrics.output_tokens if metrics else None
-            total_tokens = metrics.total_tokens if metrics else None
-            llm_calls = metrics.llm_calls if metrics else None
-            tools_used = agent_runner.last_tools_used or None
-
-            # Write session log (for all outcomes - audit trail)
-            session_path: str | None = None
-            if self._session_logger:
-                try:
-                    session_path = self._session_logger.write_session(
-                        name="heartbeat",
-                        prompt=heartbeat_prompt,
-                        response=response,
-                        tools_used=agent_runner.last_tools_used or [],
-                        metrics=agent_runner.last_metrics,
-                        duration_ms=duration_ms,
-                    )
-                except Exception as e:
-                    logger.warning(f"Failed to write heartbeat session log: {e}")
-
-            # Check for acknowledge_event tool call (takes precedence over HEARTBEAT_OK)
-            if self._ack_tool:
-                ack = self._ack_tool.get_pending_ack()
-                if ack:
-                    logger.info(
-                        f"Heartbeat acknowledged for '{self.workspace_name}': {ack.reason}"
-                    )
-                    self._record_heartbeat_event(
-                        "acknowledged",
-                        reason=ack.reason,
-                        duration_ms=duration_ms,
-                        input_tokens=input_tokens,
-                        output_tokens=output_tokens,
-                        total_tokens=total_tokens,
-                        llm_calls=llm_calls,
-                        task_count=task_count,
-                        response=response,
-                        tools_used=tools_used,
-                    )
-                    if self._token_logger and metrics:
-                        self._token_logger.log(
-                            metrics=metrics,
-                            workspace=self.workspace_name,
-                            invocation_type="heartbeat",
-                            session_key=None,
-                        )
-                    return
-
-            if self.config.suppress_ok and self._is_heartbeat_ok(response):
-                logger.debug(f"Heartbeat OK for workspace: {self.workspace_name} (suppressed)")
-                self._record_heartbeat_event(
-                    "heartbeat_ok",
-                    duration_ms=duration_ms,
-                    input_tokens=input_tokens,
-                    output_tokens=output_tokens,
-                    total_tokens=total_tokens,
-                    llm_calls=llm_calls,
-                    task_count=task_count,
-                    response=response,
-                    tools_used=tools_used,
-                )
-
-                # Log to token_usage.jsonl if available
-                if self._token_logger and metrics:
-                    self._token_logger.log(
-                        metrics=metrics,
-                        workspace=self.workspace_name,
-                        invocation_type="heartbeat",
-                        session_key=None,
-                    )
-                return
-
-            if not heartbeat_channel:
-                logger.error(
-                    f"Heartbeat channel not found: {self.config.target_channel} "
-                    f"(workspace: {self.workspace_name})"
-                )
-                self._record_heartbeat_event(
-                    "error",
-                    error="channel not found",
-                    duration_ms=duration_ms,
-                    input_tokens=input_tokens,
-                    output_tokens=output_tokens,
-                    total_tokens=total_tokens,
-                    llm_calls=llm_calls,
-                    task_count=task_count,
-                    response=response,
-                    tools_used=tools_used,
-                )
-
-                # Log to token_usage.jsonl even on error
-                if self._token_logger and metrics:
-                    self._token_logger.log(
-                        metrics=metrics,
-                        workspace=self.workspace_name,
-                        invocation_type="heartbeat",
-                        session_key=None,
-                    )
-                return
-
-            if heartbeat_session_key:
-                # Delivery routing
-                delivery = self.config.delivery
-
-                # Channel delivery ("channel" or "both")
-                if delivery in ("channel", "both"):
-                    await heartbeat_channel.send_message(session_key=heartbeat_session_key, content=response)
-                    logger.info(f"Heartbeat notification sent to {self.config.target_channel}/{heartbeat_session_key}")
-
-                # Agent queue injection ("agent" or "both")
-                if delivery in ("agent", "both") and self._result_callback and session_path:
-                    try:
-                        # Build injection content with truncation
-                        output = response
-                        if len(output) > INJECTION_TRUNCATION_LIMIT:
-                            output = output[:INJECTION_TRUNCATION_LIMIT]
-                            injection_content = HEARTBEAT_RESULT_TRUNCATED_TEMPLATE.format(
-                                output=output, session_path=session_path,
-                            )
-                        else:
-                            injection_content = HEARTBEAT_RESULT_TEMPLATE.format(
-                                output=output, session_path=session_path,
-                            )
-                        await self._result_callback(heartbeat_session_key, injection_content)
-                        logger.info(f"Heartbeat result injected into agent queue for {heartbeat_session_key}")
-                    except Exception as e:
-                        logger.warning(f"Failed to inject heartbeat result: {e}")
-
-                self._record_heartbeat_event(
-                    "ran",
-                    duration_ms=duration_ms,
-                    input_tokens=input_tokens,
-                    output_tokens=output_tokens,
-                    total_tokens=total_tokens,
-                    llm_calls=llm_calls,
-                    task_count=task_count,
-                    response=response,
-                    tools_used=tools_used,
-                )
-
-                # Log to token_usage.jsonl
-                if self._token_logger and metrics:
-                    self._token_logger.log(
-                        metrics=metrics,
-                        workspace=self.workspace_name,
-                        invocation_type="heartbeat",
-                        session_key=heartbeat_session_key,
-                    )
-            else:
-                # No routing configured (no target ID or unsupported channel)
-                delivery = self.config.delivery
-                if delivery in ("agent", "both") and self._result_callback:
-                    logger.warning(
-                        f"Heartbeat delivery configured for agent injection but no session_key available "
-                        f"(workspace: {self.workspace_name}, delivery: {delivery})"
-                    )
-                else:
-                    logger.warning(
-                        f"Heartbeat response generated but no routing configured "
-                        f"(workspace: {self.workspace_name}, response length: {len(response)})"
-                    )
-
-                self._record_heartbeat_event(
-                    "ran",
-                    reason="no routing configured",
-                    duration_ms=duration_ms,
-                    input_tokens=input_tokens,
-                    output_tokens=output_tokens,
-                    total_tokens=total_tokens,
-                    llm_calls=llm_calls,
-                    task_count=task_count,
-                    response=response,
-                    tools_used=tools_used,
-                )
-
-                # Log to token_usage.jsonl
-                if self._token_logger and metrics:
-                    self._token_logger.log(
-                        metrics=metrics,
-                        workspace=self.workspace_name,
-                        invocation_type="heartbeat",
-                        session_key=None,
-                    )
-
-        except Exception as e:
-            duration_ms = (time_module.monotonic() - start_time) * 1000
-            logger.error(
-                f"Heartbeat check failed for workspace '{self.workspace_name}': {e}",
-                exc_info=True,
-            )
-            self._record_heartbeat_event(
-                "error",
-                error=str(e),
-                duration_ms=duration_ms,
-                task_count=task_count,
-            )
-        finally:
-            clear_channel_context()
-            # Clear any stale acknowledge state so a failed run cannot bleed
-            # into the next heartbeat cycle.
-            if self._ack_tool:
-                self._ack_tool.reset()
